rename: 'configuration/objects' -> 'configuration/options'
