+++ src/configuration/options/formatting-rule.ts
